backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-extensions==3.2.3
django-simple-history==3.5.0
django-taggit==5.0.1
djangorestframework==3.15.1
sqlparse==0.5.0
